backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
sse-starlette==2.1.3
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0